Fix goroutine leaks only server (#1241)
